+++ packages/expo-modules-core/android/src/main/java/expo/modules/core/logging/Logger.kt
@@ -1,43 +1,18 @@
 package expo.modules.core.logging
 
-import android.content.Context
 import android.util.Log
 import expo.modules.BuildConfig
 
 /**
- * Native Android logging class for Expo, with options to direct logs
+ * Logging class for Expo, with options to direct logs
  * to different destinations.
  */
 class Logger(
   /**
-   * String that defines the tag used with the Android native logger, and
-   * also used to derive the file name for logging to a file
+   * LogHandler instances to which logs should be send
    */
-  category: String,
-  /**
-   * Android context is required if logging to a file
-   */
-  context: Context? = null,
-  /**
-   * One of the predefined LoggerOptions values
-   */
-  options: LoggerOptions
+  private val logHandlers: List<LogHandler>
 ) {
-
-  private val handlers = mutableListOf<LogHandler>().apply {
-    if (options.contains(LoggerOptions.logToOS)) {
-      this.add(OSLogHandler(category))
-    }
-    if (options.contains(LoggerOptions.logToFile)) {
-      this.add(
-        PersistentFileLogHandler(
-          category,
-          requireNotNull(context) { "You have to provide the `Context` to create a file logger" }
-        )
-      )
-    }
-  }.toList()
-
   private val minOSLevel = when (BuildConfig.DEBUG) {
     true -> Log.DEBUG
     else -> Log.INFO
@@ -92,9 +67,23 @@ class Logger(
     log(LogType.Fatal, message, cause)
   }
 
+  /**
+   * Starts a timer that can be used to compute the duration of an operation. Upon calling
+   * `stop` on the returned object, a timer entry will be logged.
+   */
+  fun startTimer(logFormatter: (duration: Long) -> String): LoggerTimer {
+    val start = System.currentTimeMillis()
+    return object : LoggerTimer {
+      override fun stop() {
+        val end = System.currentTimeMillis()
+        log(LogType.Timer, logFormatter(end - start))
+      }
+    }
+  }
+
   private fun log(type: LogType, message: String, cause: Throwable? = null) {
     if (LogType.toOSLogType(type) >= minOSLevel) {
-      handlers.forEach { handler ->
+      logHandlers.forEach { handler ->
         handler.log(type, message, cause)
       }
     }