+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/CoreLogger.kt
@@ -1,6 +1,6 @@
 package expo.modules.kotlin
 
+import expo.modules.core.logging.LogHandlers
 import expo.modules.core.logging.Logger
-import expo.modules.core.logging.LoggerOptions
 
-internal val logger = Logger("ExpoModulesCore", null, options = LoggerOptions.logToOS)
+internal val logger = Logger(listOf(LogHandlers.createOSLogHandler("ExpoModulesCore")))