+++ packages/expo-updates/android/src/main/java/expo/modules/updates/logging/UpdatesLogger.kt
@@ -1,18 +1,16 @@
 package expo.modules.updates.logging
 
 import android.content.Context
+import expo.modules.core.logging.LogHandlers
 import expo.modules.core.logging.LogType
 import expo.modules.core.logging.Logger
-import expo.modules.core.logging.LoggerOptions
-import java.lang.Exception
-import java.util.*
+import expo.modules.core.logging.LoggerTimer
+import java.util.Date
 
 /**
  * Class that implements logging for expo-updates with its own logcat tag
  */
-class UpdatesLogger(
-  context: Context
-) {
+class UpdatesLogger(context: Context) {
 
   fun trace(
     message: String,
@@ -27,7 +25,7 @@ class UpdatesLogger(
     updateId: String?,
     assetId: String?
   ) {
-    logger.trace(logEntryString(message, code, LogType.Trace, updateId, assetId))
+    logger.trace(logEntryString(message, code, LogType.Trace, null, updateId, assetId))
   }
 
   fun debug(
@@ -43,7 +41,7 @@ class UpdatesLogger(
     updateId: String?,
     assetId: String?
   ) {
-    logger.debug(logEntryString(message, code, LogType.Debug, updateId, assetId))
+    logger.debug(logEntryString(message, code, LogType.Debug, null, updateId, assetId))
   }
 
   fun info(
@@ -59,7 +57,7 @@ class UpdatesLogger(
     updateId: String?,
     assetId: String?
   ) {
-    logger.info(logEntryString(message, code, LogType.Info, updateId, assetId))
+    logger.info(logEntryString(message, code, LogType.Info, null, updateId, assetId))
   }
 
   fun warn(
@@ -75,7 +73,7 @@ class UpdatesLogger(
     updateId: String?,
     assetId: String?
   ) {
-    logger.warn(logEntryString(message, code, LogType.Warn, updateId, assetId))
+    logger.warn(logEntryString(message, code, LogType.Warn, null, updateId, assetId))
   }
 
   fun error(
@@ -93,7 +91,7 @@ class UpdatesLogger(
     assetId: String?,
     exception: Exception? = null
   ) {
-    logger.error(logEntryString(message, code, LogType.Error, updateId, assetId, exception))
+    logger.error(logEntryString(message, code, LogType.Error, null, updateId, assetId, exception))
   }
 
   fun fatal(
@@ -111,21 +109,29 @@ class UpdatesLogger(
     assetId: String?,
     exception: Exception? = null
   ) {
-    logger.fatal(logEntryString(message, code, LogType.Fatal, updateId, assetId, exception))
+    logger.fatal(logEntryString(message, code, LogType.Fatal, null, updateId, assetId, exception))
+  }
+
+  fun startTimer(label: String): LoggerTimer {
+    return logger.startTimer { duration ->
+      logEntryString(label, UpdatesErrorCode.None, LogType.Timer, duration, null, null, null)
+    }
   }
 
   // Private methods and fields
 
   private val logger = Logger(
-    EXPO_UPDATES_LOGGING_TAG,
-    context,
-    LoggerOptions.union(listOf(LoggerOptions.logToOS, LoggerOptions.logToFile))
+    listOf(
+      LogHandlers.createOSLogHandler(EXPO_UPDATES_LOGGING_TAG),
+      LogHandlers.createPersistentFileLogHandler(context, EXPO_UPDATES_LOGGING_TAG)
+    )
   )
 
   private fun logEntryString(
     message: String,
     code: UpdatesErrorCode,
     level: LogType,
+    duration: Long?,
     updateId: String?,
     assetId: String?,
     exception: Exception? = null
@@ -150,6 +156,7 @@ class UpdatesLogger(
       message,
       code.code,
       level.type,
+      duration,
       updateId,
       assetId,
       stacktrace